Adds debug modules to wally
This commit adds the Debug Transport Module (DTM) and the Debug Module
(DM) to wallypipelinedsoc.sv, allowing it to be enabled with
DEBUG_SUPPORTED set in the chosen configuration. Currently, the Debug
Module does not interact with the processor like it does in our
testing repository "wally-debug-test". This will be added next.