Added debugger submodule to primary testbench.
This commit adds a submodule to the primary testbench that simulates
the function of a debugger interacting with the debug module across a
jtag port. This debugger submodule will only be used if
DEBUG_SUPPORTED is set in the chosen configuration. This debugger
simulates how OpenOCD would interact with the Debug Module (DM)
through the Debug Transport Module (DTM) using the Debug Module
Interface (DMI).

The debugger submodule has the capability of loading in a file of
testvectors, generated straight from a log from OpenOCD, to test the
debug module and the debug mode capabilities of the processor while
the processor runs another assembly test. This testvector file gives
both the inputs to the debug transport module as well as the expected
outputs and verifies for each transaction that simulation behaves like
OpenOCD expects.

Future improvements:
1. A test will need to be written that runs for a long time, so that
the debugger has time to perform its actions at its lower clock speed.

2. The debugger will need to load in separate tests that test various
features of the debug specification, like dmactive or ndmreset.