Added Abstract Register Read and Write to CVW.
As of right now this commit is untested, but it passes lint and
regression. Next, the code written in this commit will be tested
against 32 bit and 64 bit architectures.

This commit adds several things:
- Abstract read and write for GPRs and CSRs
- 32 bit and 64 bit support using two Data registers in the Debug
Module.

This was accomplished by adding extra debug signals to the port
declarations of csr.sv, privilege.sv, datapath.sv, ieu.sv, and
wallypipelinedcore.sv. Additionally, a couple of minor changes were
made to csrd.sv.