Set up debug tests and fix many bugs.
Several small bugs are fixed in this commit, especially port width
bugs, or missing statements. Many OpenOCD testvectors are passing now,
except for ones where the values of CSRs are different than the
previous simpler processor design used in the `wally-debug-test`
repository.

The test tries to run the OpenOCD testvectors while Wally executes the
32 bit SPI peripheral test. Currently, the test just stops at a
certain point and stops executing and the watchdog timer is not firing
off. This will require examining waveforms in the GUI to debug.