fix(debug): fixed abstract GPR read enable pin
This fixes the Abstract Read and Write for GPRs. In Debug Mode, the
RegWrite signal for the RegFile needs to be controlled by the Debug
Module, otherwise, the pin is indefinitely high due to RegWriteW being
a control signal. A few signals and a mux was added to take care of
the problem, handing the control of the RegFile write enable to the
Debug Module. This possibly needs to be refactored in the future.