feat: testbench loads testvectors for debugger simultaneously
In this commit, several features have been added:
1. Proper test generation in the `tests/debug` directory. Now the
assembly tests are have their label files and memfiles generated on
top of the object dumps and debugger testvectors.
2. The testbench now loads and executes debugger testvectors at the
same time that it loads the assembly tests.

However, setting the PC to the DPC on resume is currently not
working. Additionally, many of the testvectors and the expected
outputs are not matching the output of Spike's OpenOCD. This will need
to be investigated.