feat: DPC sets PCF and Flushes on Resume
This commit includes a working method of setting PCF using the value
stored in DPC in Wally. This allows a debugger to set this register
and resume execution at any arbitrary point. This works with the test
WALLY-debug-02.S. Also updated the debug Makefile to detect if the
WALLY environment variable is not set.

This commit also makes the field aarsize in the debug module aware of
XLEN. This way commands that support 64 bit widths are supported only
when XLEN == 64.