fix: Implemented correct resumeack and havereset bits
The any/allresumeack bits and the anyhavereset/allhavereset bits
function as expected now. However, there appears to be a discrepancy
between how Spike does it and how my implementation does it. It's a
little confusing. Spike does not deassert the resumeack bits upon
intiating a halt request, whereas mine does. I'm not sure if that's
correct or not.