feat: Debug Module now works on FPGA
A couple of things to note about this commit:
1. Halting, Resuming, reading GPRs and CSRs are the features that have
been attempted on the FPGA. These features work during Linux
execution, but if tried during boot, it breaks the boot
process. Sometimes during Linux execution certain kernel error
messages can print out on resume but Linux will still run after the
error. All of this to say that there appears to still be states the
processor can be in that aren't accounted for when halting, causing
issues. These cases will need to be explored more thoroughly with more
rigorous tests and FPGA debugging.

2. This commit also adds an extra test to the `tests/debug`
directory. This test is capable of reading an object dump file and
setting `dpc` to the addresses corresponding to various labels. This
is useful for ensuring synchronization between what Wally executes and
what the Debugger does.