feat(debug): Halting on reset
This commit adds the ability to halt on reset using the debug
module. This has been tested with the addition of the WALLY-debug-03.S
test. Though this feature works, it needs a little bit more
work. Specifically:

1. The test needs to be redesigned somewhat. I mistakenly wrote a test
where I try to rewrite things in the ROM in Wally. In Spike, this test
passes and generates the expected memory signature, but Wally doesn't
do this. I'm thinking the test can be reworked to use the signature to
handle ensuring the correct signature region after resetting.

2. Currently, upon reset, and if HaltReq is asserted high, the Hart
does halt. However, it halts when the first instruction becomes valid
in the Memory stage. According to the debug spec, this is incorrect,
as the Instruction shouldn't have even been executed yet. Section 4.6
of the debug specification, version 1.0, gives the specific requirement.