feat(debug): Testbench runs multiple debug tests
Now the debugger file starts a fork to run the testvectors, allowing
the main fork to go back to waiting for the negative edge of the
reset. This allows interruption of the debugger testvector tasks in
order to start the next test. Because of this, multiple tests can now
run back to back along with the assembly.

Additionally, I'm uploading a couple of tests I forgot to upload
before. WALLY-debug-03.S and WALLY-debug-dmactive.S. The "dmactive"
test is badly named, as it turned into a test of resuming and halting
multiple times and writing the DPC csr to change the PC upon resume. I
will change the test later and move the current version of the test to
a file called WALLY-debug-dpc.S or something.