feat(debug): Debugger now tracks testvector state
The reason for tracking the state is to identify when we can excuse
certain differences between spike's implementation and Wally's on
certain register reads. Doing this uncovered a couple of bugs with the
Debug Module as well. So we are now at a point where at least all of
the testvectors in the test WALLY-debug-dmactive are passing. Some of
exceptions occur because of unimplemented features. Therefore, it will
be necessary to come to back and reverse those exceptions in order to
correctly see where the discrepencies are.

Next, I'll make sure the print is less verbose and only prints the
discrepencies that have not been excused so far. I will also get
Imperas working with wsim, possibly adding another argument to wsim in
order to get the simulation working.