fix(debug): Fixed resuming from current PC
The hardware was setup to change the PC in the event DPC was
changed. However, the logic did not take into account what happens if
it's not changed. Before this commit, the assumption was that if we
are resuming at a different address, we want to flush the decode and
execute stages, but this doesn't need to happen if we are resuming
from the current location. This commit fixes that by tracking if the
DPC actually changes and only flushing if it does.